% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/99_HELPERS.R
\name{algo.ui}
\alias{algo.ui}
\title{Server side for monobin functions' inputs}
\usage{
algo.ui(id)
}
\arguments{
\item{id}{Namespace id.}
}
\value{
No return value, server side call for user interface of the selected binning algorithm.
}
\description{
Server side for monobin functions' inputs
}
\examples{
if 	(interactive()) {
	algo.ui(id = "monobin")
	}
}
