% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_DESC_IMPUT.R
\name{di.server}
\alias{di.server}
\title{Descriptive statistics and imputation server side module}
\usage{
di.server(id)
}
\arguments{
\item{id}{Namespace id.}
}
\description{
Descriptive statistics and imputation server side module
}
