% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_DESC_IMPUT.R
\name{di.ui}
\alias{di.ui}
\title{Descriptive statistics and imputation user interface module}
\usage{
di.ui(id)
}
\arguments{
\item{id}{Namespace id.}
}
\description{
Descriptive statistics and imputation user interface module
}
