% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/99_SYNC.R
\name{hide.dwnl.buttons}
\alias{hide.dwnl.buttons}
\title{Hide download buttons from descriptive statistics module}
\usage{
hide.dwnl.buttons(id)
}
\arguments{
\item{id}{Namespace id.}
}
\description{
Hide download buttons from descriptive statistics module
}
