% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_MONOBIN.R
\name{mb.server}
\alias{mb.server}
\title{Monobin module - server side}
\usage{
mb.server(id)
}
\arguments{
\item{id}{Namespace id.}
}
\value{
No return value, called for monobin module server side.
}
\description{
Monobin module - server side
}
\examples{
if 	(interactive()) {
	mb.server(id = "monobin")
	}
}
