% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_MONOBIN.R
\name{mb.ui}
\alias{mb.ui}
\title{Monobin module - user interface}
\usage{
mb.ui(id)
}
\arguments{
\item{id}{Namespace id.}
}
\value{
No return value, called for monobin module user interface.
}
\description{
Monobin module - user interface
}
\examples{
if 	(interactive()) {
	mb.ui(id = "monobin")
	}
}
