% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/99_HELPERS.R
\name{mono.inputs.check}
\alias{mono.inputs.check}
\title{Check for numeric arguments - monobin module}
\usage{
mono.inputs.check(x, args.e)
}
\arguments{
\item{x}{Binning algorithm from monobin package.}

\item{args.e}{Argument elements of selected monobin function.}
}
\description{
Check for numeric arguments - monobin module
}
