% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/99_HELPERS.R
\name{monobin.fun}
\alias{monobin.fun}
\title{Evaluation expression of selected monobin function and its arguments}
\usage{
monobin.fun(x)
}
\arguments{
\item{x}{Binning algorithm from monobin package.}
}
\description{
Evaluation expression of selected monobin function and its arguments
}
