% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/99_HELPERS.R
\name{monobin.run}
\alias{monobin.run}
\title{Run monobin algorithm for selected inputs}
\usage{
monobin.run(algo, target.n, rf, sc, args.e, db)
}
\arguments{
\item{algo}{Binning algorithm from monobin package.}

\item{target.n}{Selected target.}

\item{rf}{Vector of selected numeric risk factors.}

\item{sc}{Numeric vector of special case values.}

\item{args.e}{Argument elements of selected monobin function.}

\item{db}{Data frame of target and numeric risk factors.}
}
\description{
Run monobin algorithm for selected inputs
}
