% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monobinShinyApp.R
\name{monobinShinyApp}
\alias{monobinShinyApp}
\title{Start shiny UI for monobin package}
\usage{
monobinShinyApp()
}
\description{
Start shiny UI for monobin package
}
\examples{
if 	(interactive()) {
	suppressMessages(library(monobinShiny))	
	monobinShinyApp()
	}
}
