% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/99_HELPERS.R
\name{num.inputs}
\alias{num.inputs}
\title{Numeric arguments - monobin module}
\usage{
num.inputs(x)
}
\arguments{
\item{x}{Binning algorithm from monobin package.}
}
\description{
Numeric arguments - monobin module
}
