% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/99_HELPERS.R
\name{out.impute}
\alias{out.impute}
\title{Outliers - imputation}
\usage{
out.impute(tbl, rf, ub, lb, sc)
}
\arguments{
\item{tbl}{Data frame with risk factors ready for imputation.}

\item{rf}{Vector of risk factors to be imputed.}

\item{ub}{Upper bound percentiles.}

\item{lb}{Lower bound percentiles.}

\item{sc}{Numeric vector of special case values.}
}
\description{
Outliers - imputation
}
