% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/99_HELPERS.R
\name{pct.ui}
\alias{pct.ui}
\title{pct.bin - monobin functions' inputs}
\usage{
pct.ui(id)
}
\arguments{
\item{id}{Namespace id.}
}
\value{
No return value, called for user interface of the pct.bin - monobin functions' inputs.
}
\description{
pct.bin - monobin functions' inputs
}
\examples{
if 	(interactive()) {
	output$algo.args <- renderUI({tagList(switch(algo.select, "cum.bin" = cum.ui(id = id),
							  "iso.bin" = iso.ui(id = id),
							  "ndr.bin" = ndr.sts.ui(id = id),
							  "sts.bin" = ndr.sts.ui(id = id),
							  "pct.bin" = pct.ui(id = id),
							  "woe.bin" = woe.ui(id = id),
						          "mdt.bin" = mdt.ui(id = id)))
		})	
	}
}
