% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/99_HELPERS.R
\name{sc.check}
\alias{sc.check}
\title{Special cases - check input values}
\usage{
sc.check(x)
}
\arguments{
\item{x}{Numeric vector of special case values.}
}
\description{
Special cases - check input values
}
