% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/99_SYNC.R
\name{sync.m23.imp}
\alias{sync.m23.imp}
\title{Sync between descriptive statistics and monobin module after imputation process}
\usage{
sync.m23.imp(id, num.rf, module)
}
\arguments{
\item{id}{Namespace id.}

\item{num.rf}{Vector of updated numeric risk factors.}

\item{module}{Descriptive statistic or monobin module.}
}
\description{
Sync between descriptive statistics and monobin module after imputation process
}
