% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/99_SYNC.R
\name{upd.dm}
\alias{upd.dm}
\title{Update data manager output}
\usage{
upd.dm(id, dummy)
}
\arguments{
\item{id}{Namespace id.}

\item{dummy}{A logical value indicating whether gcd data (from monobin package) or specific csv file are imported.}
}
\description{
Update data manager output
}
