% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/99_SYNC.R
\name{upd.si.m23}
\alias{upd.si.m23}
\title{Update sync between descriptive statistics and monobin module}
\usage{
upd.si.m23(upd.rf, num.rf, session)
}
\arguments{
\item{upd.rf}{Vector of risk factor field ids that need to be updated.}

\item{num.rf}{Vector of updated numeric risk factors.}

\item{session}{Session object .}
}
\description{
Update sync between descriptive statistics and monobin module
}
