% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/99_HELPERS.R
\name{check.vars}
\alias{check.vars}
\title{Check for categorical variables when importing the data}
\usage{
check.vars(tbl)
}
\arguments{
\item{tbl}{Imported data frame.}
}
\value{
Returns a character vector which describes variables type of imported data frame.
}
\description{
Check for categorical variables when importing the data
}
\examples{
if 	(interactive()) {
	check.msg <- check.vars(tbl = rv$db)
	}
}
