% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/99_HELPERS.R
\name{monobin.fun}
\alias{monobin.fun}
\title{Evaluation expression of the selected monobin function and its arguments}
\usage{
monobin.fun(x)
}
\arguments{
\item{x}{Binning algorithm from monobin package.}
}
\value{
Returns an evaluation expression of the selected monobin algorithm.
}
\description{
Evaluation expression of the selected monobin function and its arguments
}
\examples{
if 	(interactive()) {
	expr.eval <- monobin.fun(x = algo)
	}
monobin.fun(x = "ndr.bin")


}
