% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monobinShinyApp.R
\name{monobinShinyApp}
\alias{monobinShinyApp}
\title{Starts shiny application for the monobin package}
\usage{
monobinShinyApp()
}
\value{
Starts shiny application for the monobin package.
}
\description{
Starts shiny application for the monobin package
}
\examples{
if 	(interactive()) {
	suppressMessages(library(monobinShiny))	
	monobinShinyApp()
	}
}
