% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/99_HELPERS.R
\name{num.inputs}
\alias{num.inputs}
\title{Numeric arguments - monobin module}
\usage{
num.inputs(x)
}
\arguments{
\item{x}{Binning algorithm from monobin package.}
}
\value{
Returns a list of two vectors: index and UI element label of numeric arguments of
the selected monobin function.
}
\description{
Numeric arguments - monobin module
}
\examples{
if 	(interactive()) {
	inp.indx <- num.inputs(x = x)
	}
num.inputs(x = "cum.bin")

}
