% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/99_SYNC.R
\name{sync.m23}
\alias{sync.m23}
\title{Sync between descriptive statistics and monobin module after data import}
\usage{
sync.m23(id, num.rf, module)
}
\arguments{
\item{id}{Namespace id.}

\item{num.rf}{Vector of updated numeric risk factors.}

\item{module}{Descriptive statistic or monobin module.}
}
\value{
No return value, called in order to sync between descriptive statistics and monobin modules' UI elements after data import.
}
\description{
Sync between descriptive statistics and monobin module after data import
}
\examples{
if 	(interactive()) {
observeEvent(rv$sync, {
	sync.m23(id = "desc.imputation", 
		 num.rf = rv$num.rf,
		 module = "desc")
	sync.m23(id = "monobin", 
		 num.rf = rv$num.rf,
		 module = "monobin")
	rv$rf.imp <- NULL
	rv$rf.out <- NULL
	}, ignoreInit = TRUE)
	}
}
